+++ packages/dx-react-chart/src/types/index.ts
@@ -37,6 +37,7 @@ export * from './plugins.title.types';
 export * from './plugins.tooltip.types';
 export * from './plugins.zoom-pan.types';
 export * from './utils.series-helpers.types';
+export * from './utils.with-animation.types';
 export * from './templates.axis.types';
 export * from './templates.core.types';
 export * from './chart.types';