+++ packages/dx-react-grid-demos/src/demo-sources/core-action/bootstrap3/simple-action.jsx
@@ -72,7 +72,7 @@ const FilterPanel = () => (
           <div>
             Filter: (
             {JSON.stringify(filter)}
-)
+            )
             {' '}
             <button
               type="button"