+++ packages/dx-react-grid-material-ui/src/templates/table-group-cell/content.jsx
@@ -18,7 +18,7 @@ const ContentBase = ({
   >
     <strong>
       {column.title || column.name}
-        :
+      :
       {' '}
     </strong>
     {children || String(row.value)}